requirements.txt
# App runtime
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.6
openai>=1.0.0

# Testing
pytest>=7.4.0

# Scripts / optional
requests>=2.28.0
